Reorganize installation scripts
- Just run `./INSTALL`
- Options can be stored in `INSTALL.opts`
- The main logic lives in creusot-install
