+++ .github/workflows/deploy.yml
@@ -40,8 +40,7 @@ jobs:
         key: ${{ runner.os }}-cargo-${{ hashFiles('**/Cargo.lock') }}
     - name: Install Creusot
       run: |
-        cargo install --path cargo-creusot
-        cargo creusot setup install --only-creusot-rustc
+        ./INSTALL --skip-extra-tools
     - name: Dummy creusot setup
       run: |
         mkdir -p ~/.config/creusot
