+++ .github/workflows/rust.yml
@@ -130,11 +130,11 @@ jobs:
       run: |
         tar -xzf _opam.tar.gz
         echo /home/runner/work/creusot/creusot/_opam/bin >> $GITHUB_PATH
-    - name: run cargo creusot setup install
+    - name: Install solvers
       if: steps.cache-creusot-setup.outputs.cache-hit != 'true'
       # Use only 2 parallel provers, because more provers (4) makes replaying proofs unstable
       run: |
-        cargo run --bin cargo-creusot creusot setup install --provers-parallelism 2
+        ./INSTALL --skip-cargo-creusot --skip-creusot-rustc --provers-parallelism 2
         echo -e "\n>> ~/.config/creusot/Config.toml:\n"
         cat ~/.config/creusot/Config.toml
         echo -e "\n>> ~/.config/creusot/why3.conf:\n"
@@ -152,18 +152,6 @@ jobs:
           ~/.cargo/git
           target
         key: ${{ runner.os }}-cargo-install-${{ hashFiles('**/Cargo.lock') }}
-    - uses: actions/cache@v4
-      id: cache-creusot-setup
-      with:
-        path: |
-          ~/.config/creusot
-          ~/.local/share/creusot
-          _opam/lib/why3find/packages
-        key: ${{ runner.os }}-cargo-creusot-setup-${{ hashFiles('creusot-deps.opam', 'creusot-setup/src/tools_versions_urls.rs', 'creusot-setup/src/config.rs') }}
-    - name: Install
-      run: |
-        cargo install --path cargo-creusot
-        cargo creusot setup install --only-creusot-rustc
     - name: download why3 and why3find
       uses: actions/download-artifact@v4
       with:
@@ -171,11 +159,10 @@ jobs:
     - name: unpack why3 and why3find
       run: tar -xzf _opam.tar.gz
     - name: Setup
+      # Add /home/runner/work/creusot/creusot/_opam/bin to PATH just for this step
       run: |
-        # Add /home/runner/work/creusot/creusot/_opam/bin to PATH just for this step
         export PATH=/home/runner/work/creusot/creusot/_opam/bin:$PATH
-        cargo creusot setup install --skip-creusot-rustc
-      if: steps.cache-creusot-setup.outputs.cache-hit != 'true'
+        ./INSTALL
     - name: test cargo creusot new
       run: |
         set -x
