+++ .gitignore
@@ -8,6 +8,7 @@ _opam
 
 # Creusot local config for development
 /.creusot-config
+INSTALL.opts
 
 # Creusot files
 *.creusot
