+++ Cargo.toml
@@ -9,6 +9,7 @@ members = [
   "creusot-contracts-dummy",
   "creusot-metadata",
   "creusot-dev-config",
+  "creusot-install",
   "why3",
   "why3tests",
   "pearlite-syn",
