+++ HACKING.md
@@ -7,12 +7,12 @@ working on the Creusot codebase.
 
 The testsuite will use the global Creusot configuration managed by 
 `cargo creusot setup`. 
-You first need to have successfully run `cargo creusot setup install` as
+You first need to have successfully run `./INSTALL` as
 detailed in the README installation instructions.
 
 **To be able to use custom versions of Why3 or the solvers** (instead of the
 built-in ones expected by Creusot), one can pass extra flags to 
-`cargo creusot setup install` (see also `--help`):
+`./INSTALL` (see also `--help`):
 - `--external <TOOL>` to specify that a solver should be looked up from the path
 - `--no-check-version <TOOL>` to allow unexpected versions of a given tool
 
@@ -54,7 +54,7 @@ If the proof of a test is broken (e.g.
 ## Calling why3 (and why3_tools, etc): shell environment setup
 
 To invoke why3 (manually or in scripts) with the same binary/configuration as
-setup by `cargo creusot setup`, one needs to setup a shell environment with the
+setup by `./INSTALL`, one needs to setup a shell environment with the
 correct PATH and environment variables.
 
 To do so, we provide the following command:
@@ -63,7 +63,7 @@ eval $(cargo run --bin dev-env)
 ```
 
 After that, the `why3` binary in PATH will be the one configured by
-`cargo creusot setup`, using the adequate configuration file (through the
+`./INSTALL`, using the adequate configuration file (through the
 `WHY3CONFIG` environment variable).
 
 ## Upgrading Why3
@@ -79,7 +79,7 @@ of the file (URLs and `git-XXXX` versions), and the `git-XXXX` versions in the
 
 - Install why3-tools: `opam pin git+https://github.com/xldenis/why3-tools`
 - Install the newer prover, make it available in `$PATH`
-- Setup Creusot to use it: `cargo creusot setup install --no-check-version <PROVER> --external <PROVER>`
+- Setup Creusot to use it: `./INSTALL --no-check-version <PROVER> --external <PROVER>`
 - Run `eval $(cargo run --bin dev-env)`
 - Use the `./testsuite_upgrade_prover` script to update why3 sessions in the testsuite.
   Launch the script without arguments to have some usage instructions.
