+++ INSTALL
@@ -1,0 +1,13 @@
+#!/bin/sh
+# Install Creusot
+# Additional arguments can be stored in a file `INSTALL.opts`. For example:
+#     $ cat "--external z3" > INSTALL.opts
+#     $ ./INSTALL
+# To see a list of possible options, `./INSTALL --help`.
+if [ -f INSTALL.opts ] ; then
+    ARGS=$(cat INSTALL.opts)
+else
+    ARGS=""
+fi
+set -xe
+cargo run --bin creusot-install -- $ARGS $@
