+++ creusot-install/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "creusot-install"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+clap = { version = "4.5", features = ["derive", "env"] }
+creusot-setup = {path = "../creusot-setup"}
+anyhow = "1.0"
+indoc = "2.0.5"
+reqwest = { version = "0.12", features = ["blocking"] }
+sha2 = "0.10"
+hex = "0.4"
+zip = "2.2"
