+++ creusot-setup/Cargo.toml
@@ -11,9 +11,3 @@ toml = { version = "0.8", features = ["preserve_order"] }
 directories = "6.0"
 which = "7.0"
 anyhow = "1.0"
-reqwest = { version = "0.12", features = ["blocking"] }
-zip = "2.2"
-hex = "0.4"
-sha2 = "0.10"
-indoc = "2.0"
-
